Refactored Files

+++ Notes&Docs/Views/FileView.swift
@@ -10,14 +10,17 @@ import SwiftUI
 
 struct FileView: View {
     @Environment(\.presentationMode) private var presentationMode: Binding<PresentationMode>
-    @EnvironmentObject var renameAlert: TextAlert
-    @ObservedObject var doc: Document
-    @EnvironmentObject var master: MasterDirectory
     @Environment(\.editMode) var editMode
+    @EnvironmentObject var renameAlert: TextAlert
+    @EnvironmentObject var master: MasterDirectory
+    
+    @ObservedObject var doc: Document
+    @Binding private var selection: Set<Document>
     @State private var viewSelected = false
     @State private var presentView = false
-    @Binding private var selection: Set<Document>
     @State private var thumbnail: Image = Image(systemName: "pencil")
+    
+    private var isEditing: Bool { editMode?.wrappedValue == .active }
     
     init(_ doc: Document, _ selection: Binding<Set<Document>>) {
         self.doc = doc
@@ -28,14 +31,11 @@ struct FileView: View {
         ZStack {
             Button {
                 viewSelected.toggle()
-                if editMode?.wrappedValue == .inactive {
+                if !isEditing {
                     presentView.toggle()
-                } else if editMode?.wrappedValue == .active {
-                    if selection.contains(doc) {
-                        selection.remove(doc)
-                    } else {
-                        selection.update(with: doc)
-                    }
+                } else {
+                    guard selection.contains(doc) else { selection.update(with: doc); return; }
+                    selection.remove(doc)
                 }
             } label: {
                 VStack {
@@ -45,34 +45,27 @@ struct FileView: View {
                             .scaleEffect(0.90)
                             .background(Color(uiColor: .systemGray6))
                             .clipShape(RoundedRectangle(cornerRadius: 15.0))
-                            
                         
                         Circle()
                             .size(CGSize(width: 20.0, height: 20.0))
-                            .foregroundColor((viewSelected && editMode?.wrappedValue == .active) ? .accentColor : .clear)
+                            .foregroundColor((viewSelected && isEditing) ? .accentColor : .clear)
                             .frame(width: 20.0, height: 20.0)
                             .overlay {
                                 ZStack {
                                     Image(systemName: "checkmark")
                                         .resizable()
                                         .frame(width: 10, height: 10, alignment: .center)
-                                        .opacity((viewSelected && editMode?.wrappedValue == .active) ? 1 : 0)
+                                        .opacity((viewSelected && isEditing) ? 1 : 0)
                                         .foregroundColor(.white)
                                     Circle()
-                                        .stroke((viewSelected && editMode?.wrappedValue == .active) ? .white : .gray, lineWidth: 2)
+                                        .stroke((viewSelected && isEditing) ? .white : .gray, lineWidth: 2)
                                  }
-                                
                             }
                             .position(x: 126, y: 180)
-                            .opacity(editMode?.wrappedValue == .active ? 1 : 0)
+                            .opacity(isEditing ? 1 : 0)
                             
                     }
-                    .overlay {
-                        if editMode?.wrappedValue == .active {
-                            RoundedRectangle(cornerRadius: 16)
-                                .stroke(Color.accentColor, lineWidth: (viewSelected && editMode?.wrappedValue == .active) ? 2 : 0)
-                        }
-                    }
+                    .overlay { RoundedRectangle(cornerRadius: 16).stroke(Color.accentColor, lineWidth: (viewSelected && isEditing) ? 2 : 0) }
                     
                     Spacer()
                     
@@ -82,44 +75,22 @@ struct FileView: View {
                     Spacer()
                 }
                 .contextMenu {
-                    Button {
-                        // rename document
-                        
-                    } label: {
-                        Label("Rename", systemImage: "character.cursor.ibeam")
-                    }
-                    Button {
-                        // move document
-                    } label: {
-                        Label("Move", systemImage: "rectangle.portrait.and.arrow.right")
-                    }
-                    Button(role: .destructive) {
-                        doc.delete()
-                    } label: {
-                        Label("Delete Document", systemImage: "trash")
-                    }
+                    Button { } label: { Label("Rename", systemImage: "character.cursor.ibeam") }
+                    Button { } label: { Label("Move", systemImage: "rectangle.portrait.and.arrow.right") }
+                    Button(role: .destructive) { doc.delete() } label: { Label("Delete Document", systemImage: "trash") }
                 }
-                
-
             }
             
             NavigationLink(isActive: $presentView) {
                 if doc is Note {
                     NoteView(note: doc as! Note)
-                } else if doc is WordPad {
+                } else {
                     WordPadView(wordPad: doc as! WordPad)
                 }
-            } label: {
-                EmptyView()
-            }
+            } label: { EmptyView() }
         }
-        .onAppear {
-            if let data = doc.thumbnailData { thumbnail = Image(uiImage: UIImage(data: data)!)}
-        }
-        .onChange(of: editMode?.wrappedValue == .inactive) { _ in
-            selection.removeAll()
-            viewSelected = false
-        }
+        .onAppear { if let data = doc.thumbnailData { thumbnail = Image(uiImage: UIImage(data: data)!) } }
+        .onChange(of: isEditing) { _ in selection.removeAll(); viewSelected = false }
         .frame(width: 250, height: 250, alignment: .center)
         .padding()
     }
